requirements.txt
langchain==0.3.27
langchain_google_genai==2.1.10
python_dotenv==1.1.1
langchain_openai==0.3.32
langchain_groq==0.3.7
pydantic==2.11.7
langchain_community==0.3.29
pypdf==6.0.0
bs4==0.0.2
langchain_chroma
streamlit==1.49.1
wikipedia
